+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## TBD - 0.1.2
 
+- Improve query cache handling
+
 ## 2026-05-26 - 0.1.1
 
 - Add GitHub actions example and instructions to validate geofeed using CLI hook command